cmd/trayscale: refactor systray wrappers into their own file
