+++ cmd/trayscale/app.go
@@ -661,25 +661,3 @@ func (a *App) newPeerPage(peer *ipnstate.PeerStatus) *peerPage {
 
 	return &page
 }
-
-var systrayExit = make(chan func(), 1)
-
-func startSystray(onStart func()) {
-	start, stop := systray.RunWithExternalLoop(onStart, nil)
-	select {
-	case f := <-systrayExit:
-		f()
-	default:
-	}
-
-	start()
-	systrayExit <- stop
-}
-
-func stopSystray() {
-	select {
-	case f := <-systrayExit:
-		f()
-	default:
-	}
-}
